backend/requirements.txt
fastapi
uvicorn
openai
asyncpg
pydantic
python-dotenv
